fix: HttpClientsConfiguration and ReactiveHttpClientsConfiguration classes are added with limited functionality. Version is changed to 1.2.82-SNAPSHOT.
