+++ atp-auth-impl/src/main/java/org/qubership/atp/auth/springbootstarter/security/oauth2/client/config/annotation/EnableM2MRestTemplate.java
@@ -21,7 +21,11 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import org.qubership.atp.auth.springbootstarter.config.HttpClientsConfiguration;
+import org.springframework.context.annotation.Import;
+
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
+@Import({HttpClientsConfiguration.class})
 public @interface EnableM2MRestTemplate {
 }
