+++ atp-auth-impl/src/main/java/org/qubership/atp/auth/springbootstarter/security/oauth2/client/config/annotation/EnableOauth2FeignClientInterceptor.java
@@ -21,8 +21,12 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import org.qubership.atp.auth.springbootstarter.config.HttpClientsConfiguration;
+import org.springframework.context.annotation.Import;
+
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
+@Import({HttpClientsConfiguration.class})
 public @interface EnableOauth2FeignClientInterceptor {
 
 }
