+++ pom.xml
@@ -10,7 +10,7 @@
 
     <groupId>org.qubership.atp.auth</groupId>
     <artifactId>atp-auth-spring-boot-starter</artifactId>
-    <version>1.2.81-SNAPSHOT</version>
+    <version>1.2.82-SNAPSHOT</version>
     <packaging>pom</packaging>
     <modules>
         <module>atp-auth-models</module>
